pyproject.toml
[project]
name = "amail"
version = "0.1.0"
description = "Aros mail system"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]==0.115.0",
    "httpx==0.27.0",
    "jinja2==3.1.4",
    "passlib[bcrypt]>=1.7.4",
    "pydantic[email]==2.8.2",
    "pytest==8.3.3",
    "pytest-asyncio==0.24.0",
    "python-dotenv==1.0.1",
    "python-jose[cryptography]>=3.3.0",
    "resend==2.23.0",
    "ruff==0.6.9",
    "structlog==24.4.0",
    "tenacity==8.3.0",
    "uvicorn==0.30.6",
]
